I added a user review system.
Fix some bug. Added a user review system. It just a prototype and need
to be put in a template. I'm planning to add a product field so the
user can put what type of product they are talking about. There are
also a ratings and comment. We just need to workout the format and bugs.